server/requirements.txt
aniso8601==8.0.0
appdirs==1.4.4
attrs==20.1.0
branca==0.4.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
cycler==0.10.0
distlib==0.3.1
filelock==3.0.12
Fiona==1.8.13.post1
Flask==1.1.2
Flask-Cors==3.0.8
flask-marshmallow==0.15.0
Flask-RESTful==0.3.8
folium==0.11.0
geographiclib==1.50
geopandas==0.8.1
geopy==2.0.0
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
marshmallow==3.19.0
matplotlib==3.3.1
munch==2.5.0
numpy==1.19.1
packaging==24.0
pandas==1.1.1
Pillow==7.2.0
pipenv==2020.8.13
psycopg2-binary==2.9.9
pygame==1.9.6
pygeos==0.7.1
pyparsing==2.4.7
pyproj==2.6.1.post1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
Shapely==1.7.1
six==1.15.0
Unidecode==1.1.1
urllib3==1.25.10
uWSGI==2.0.30
virtualenv==20.0.31
virtualenv-clone==0.5.4
Werkzeug==1.0.1
zipp==3.1.0
